+++ HISTORY.md
@@ -2,8 +2,21 @@
 This document outlines all notable changes to this project thus far. Inspired by https://keepachangelog.com/en/1.0.0/.
 
 ## [Unreleased]
-- Input parsing: hours works (0-99), minutes + seconds in progress (0-59)
-- Alarm when timer is done
+- Custom alarm sounds
+- Upgraded UI
+- More?
+
+## [0.1.1] - 2022-10-31
+### Added
+- Alarm sound when timer is done
+- Error text for invalid input
+- requirements.txt and .gitignore
+
+### Changed
+- UI slightly modified for time input dialog
+
+### Fixed
+- Input parsing should be working as intended
 
 ## [0.1.0] - 2022-10-29
 ### Added